requirements.txt
requests==2.28.2
python-telegram-bot==13.0
python-dotenv==0.21.1